+++ DNAValidator/src/main/scala/software/sigma/nucleobase/entity/Rate.scala
@@ -1,3 +1,3 @@
 package software.sigma.nucleobase.entity
 
-case class Rate(rate: Double)
+case class Rate(rate: BigDecimal)
